+++ libsrc/ffdec_lib/test/com/jpexs/decompiler/flash/ActionScript2Test.java
@@ -1978,7 +1978,7 @@ public class ActionScript2Test extends ActionScript2TestBase {
                 + "{\r\n"
                 + "trace(_loc1_);\r\n"
                 + "}\r\n"
-                + "_loc3_.r1 = _loc2_ + 1 + \". \" + (!_loc4_ ? _loc3_.r2 = v1[_loc2_][0] : \"unk\");\r\n"
+                + "_loc3_.r1 = _loc2_ + 1 + \". \" + (!_loc4_ ? (_loc3_.r2 = v1[_loc2_][0]) : \"unk\");\r\n"
                 + "}\r\n"
                 + "trace(\"chainedAfterForInTest\");\r\n"
                 + "var v1 = {};\r\n"