+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - [#1572] SVG Export - clipping must not use groups
 - [#270] AS decompilation - switch in loop
 - [#270] AS decompilation - loop followed by try
+- [#270] AS decompilation - comma in ternar
 
 ## [14.2.0] - 2021-03-12
 ### Added